@@ -19,7 +19,7 @@ function PrivateRoute({ children, isAuth, ...rest}) {
 
 function App() {
   // We houden in de state bij of iemand is "ingelogd" (simpele versie)
-  const [isAuthenticated, toggleIsAuthenticated ] = useState(false);
+  const [isAuthenticated, toggleIsAuthenticated] = useState(false);
 
   return (
     <div>
